+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/definitions/PayByBankDefinition.kt
@@ -1,0 +1,36 @@
+package com.stripe.android.lpmfoundations.paymentmethod.definitions
+
+import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
+import com.stripe.android.lpmfoundations.paymentmethod.AddPaymentMethodRequirement
+import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodDefinition
+import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
+import com.stripe.android.lpmfoundations.paymentmethod.UiDefinitionFactory
+import com.stripe.android.model.PaymentMethod
+import com.stripe.android.ui.core.R
+
+internal object PayByBankDefinition : PaymentMethodDefinition {
+    override val type: PaymentMethod.Type = PaymentMethod.Type.PayByBank
+
+    override val supportedAsSavedPaymentMethod: Boolean = false
+
+    override fun requirementsToBeUsedAsNewPaymentMethod(
+        hasIntentToSetup: Boolean
+    ): Set<AddPaymentMethodRequirement> = setOf(
+        AddPaymentMethodRequirement.UnsupportedForSetup,
+    )
+
+    override fun requiresMandate(metadata: PaymentMethodMetadata): Boolean = false
+
+    override fun uiDefinitionFactory(
+        metadata: PaymentMethodMetadata
+    ): UiDefinitionFactory = PayByBankUiDefinitionFactory
+}
+
+private object PayByBankUiDefinitionFactory : UiDefinitionFactory.Simple() {
+    override fun createSupportedPaymentMethod(metadata: PaymentMethodMetadata) = SupportedPaymentMethod(
+        paymentMethodDefinition = PayByBankDefinition,
+        displayNameResource = R.string.stripe_paymentsheet_payment_method_pay_by_bank,
+        iconResource = R.drawable.stripe_ic_paymentsheet_pm_pay_by_bank,
+        iconResourceNight = null,
+    )
+}